Feat: Use YAML instead of ini and add run subcmd
This commit adds a run subcommand which allows users to supply commands that
they want to run in a specific environment. It also switches the supported
dovetail declaration format to be yaml as it provides an easier interface to
supply a list of commands to execute. If one command fails, the subsequent ones
don't execute

The repo's dovetail.yaml now has a dev.push environment which contains the
commands I would like executed when I want to push to GitHub.