Feat: Add release subcmd
This commit refactors the run subcmd. In particular, the logic that recursively
depth-first-searched dependencies was extracted to a new helpers module.
Additionally, the logic that determined the commands in each dependency was
migrated to the same new module. The emulation of shell execution emulation now
is also defined as a public method.

The refactor allowed for the creation of a release subcommand which appropriates
these commands to achieve its own goals. As this is also essentially a
task-running subcmd, it's not imperatively necessary. Users could have defined
their own release environments - and still can. However, this is provided for
convenience such that various release platforms can be defined centrally to an
environment.