[Packaging]: Fix Debian Packaging Rules
Fixing issues with the debian packaging such as the command to run, issues with
the toml, the automations in dovetail.yaml, and updating the changelog.

Also modified the dependent environment executor to not print "No commands to
run" since it was outputting to terminal when environments had no dependent
environments.